Move main application module to core-common repository
