--- activiti-application/src/main/java/org/activiti/application/ApplicationDiscovery.java
@@ -1,49 +1,0 @@
-/*
- * Copyright 2018 Alfresco, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.activiti.application;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
-import org.springframework.core.io.Resource;
-import org.springframework.core.io.support.ResourcePatternResolver;
-
-public class ApplicationDiscovery {
-
-    private static final String APPLICATIONS_LOCATION = "classpath:/applications/";
-    private ResourcePatternResolver resourceLoader;
-
-    public ApplicationDiscovery(ResourcePatternResolver resourceLoader) {
-        this.resourceLoader = resourceLoader;
-    }
-
-    public List<Resource> discoverApplications() {
-        List<Resource> resources = new ArrayList<>();
-        Resource resource = resourceLoader.getResource(APPLICATIONS_LOCATION);
-        if (resource.exists()) {
-            try {
-                resources = Arrays.asList(resourceLoader.getResources(APPLICATIONS_LOCATION + "**.zip"));
-            } catch (IOException e) {
-                //fixme
-                e.printStackTrace();
-            }
-        }
-        return resources;
-    }
-}
