--- activiti-application/src/main/java/org/activiti/application/ApplicationEntry.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2018 Alfresco, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.activiti.application;
-
-public class ApplicationEntry {
-
-    private String type;
-
-    private FileContent fileContent;
-
-    public ApplicationEntry(String type,
-                            FileContent fileContent) {
-        this.type = type;
-        this.fileContent = fileContent;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public FileContent getFileContent() {
-        return fileContent;
-    }
-
-}
