+++ activiti-spring-app-process/src/main/java/org/activiti/application/deployer/ProcessEntryDeployer.java
@@ -20,7 +20,7 @@ import java.util.List;
 
 import org.activiti.application.ApplicationContent;
 import org.activiti.application.FileContent;
-import org.activiti.application.ProcessEntryDiscovery;
+import org.activiti.application.discovery.ProcessEntryDiscovery;
 import org.activiti.engine.RepositoryService;
 import org.activiti.engine.repository.DeploymentBuilder;
