+++ activiti-spring-app-process/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,4 @@
 # Activiti auto-configurations
 
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-    org.activiti.application.conf.ApplicationAutoConfiguration
+    org.activiti.application.conf.ApplicationProcessAutoConfiguration
