+++ activiti-spring-app-process/src/test/java/org/activiti/application/deployer/ProcessEntryDeployerTest.java
