+++ activiti-spring-boot-starter/pom.xml
@@ -12,7 +12,7 @@
     <dependencies>
         <dependency>
             <groupId>org.activiti</groupId>
-            <artifactId>activiti-application</artifactId>
+            <artifactId>activiti-spring-app-process</artifactId>
         </dependency>
         <dependency>
             <groupId>org.activiti.core.common</groupId>
